Moves blockClass to modules folder
